rely on compiler command line for __int128 and std::decimal::decimal128
previously the availability of the types was deduced by the used compiler:
MSVC was assumed to miss them, any other compiler was assumed to have
them.

This mostly works, except on non-mainstream architectures like 32-bit arm
or mips.

Example build failures:

https://buildd.debian.org/status/fetch.php?pkg=flamerobin&arch=arm64&ver=0.9.3.12-1&stamp=1663233656&raw=0
https://buildd.debian.org/status/fetch.php?pkg=flamerobin&arch=armhf&ver=0.9.3.12-1&stamp=1663233737&raw=0
https://buildd.debian.org/status/fetch.php?pkg=flamerobin&arch=armhf&ver=0.9.3.12-2&stamp=1663429893&raw=0

The proposed solution is to rely on the tests in CMakeLists.txt for each
type. The result is passing builds on 14 Debian architectures. (all
architectures where wx3.2 is available)